Support more complex tool input schemas
The current InputSchema struct is not capable of representing nested
fields. This uses https://github.com/invopop/jsonschema to allow for full json
schemas to be generated.

If we wanted to be less opinionated and reduce dependencies, we could
follow the lead of go-openai and just type `InputSchema` as
`interface{}` instead, and json.Marshal whatever is provided